template-tanstack-start-authkit: move callback URL
