+++ template-tanstack-start-authkit/.env.local.example
@@ -2,7 +2,7 @@
 WORKOS_CLIENT_ID=client_your_client_id_here
 WORKOS_API_KEY=sk_test_your_api_key_here
 WORKOS_COOKIE_PASSWORD=your_secure_password_here_must_be_at_least_32_characters_long
-WORKOS_REDIRECT_URI=http://localhost:3000/api/auth/callback
+WORKOS_REDIRECT_URI=http://localhost:3000/callback
 
 # Convex Configuration
 VITE_CONVEX_URL=https://your-convex-deployment.convex.cloud
