+++ template-tanstack-start-authkit/src/routes/callback.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute } from '@tanstack/react-router';
 import { handleCallbackRoute } from '@workos/authkit-tanstack-react-start';
 
-export const Route = createFileRoute('/api/auth/callback')({
+export const Route = createFileRoute('/callback')({
   server: {
     handlers: {
       GET: handleCallbackRoute,
